农民删除功能实现，暂未完善

+++ src/main/java/com/zl/service/LogService.java
@@ -42,6 +42,6 @@ public interface LogService {
     * @Author: ZhuLin
     * @Date: 2019/1/10 
     */ 
-    int getLogCount();
+    Long getLogCount();
 
 }
